requirements.txt
numpy
pandas
opencv-python
face-recognition
requests
flask
PassportEye == 2.2.1
pillow 
pytesseract == 0.3.7
keras-ocr == 0.9.3